Relocate some logger variants
